backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
mammoth==1.8.0
openpyxl==3.1.2
pandas==2.2.1
psycopg2-binary==2.9.9
python-docx==1.1.0
pytz==2024.1
sqlparse==0.5.0
dj-database-url==2.1.0
whitenoise==6.6.0
Pillow==10.4.0
lxml==5.2.1
